+++ lexer/scanner.go
@@ -10,7 +10,6 @@ var keywordTags = []Tag{
 	TagEcho,
 	TagTrue,
 	TagFalse,
-	TagNothing,
 	TagEquals,
 	TagDifferent,
 	TagBigger,